Check token on every request
The token check is always run. The data inside the token payload is always passed to the load functions if the data is present. Otherwise null is passed. If the user tries to access a protected route (starting with /me) without a valid token, the user will be redirected to the /login page.